Bump third_party/protobuf to v3.4.1
This version fixes warnings when compiling on macOS.

It is generally discouraged to use the bundled version of protobuf and
rather use the system installed version. CMake, by default, will use
the system installed version. When cross compiling Caffe2 (e.g. for
mobile), it needs to use the bundled version of protobuf.