add sanity check to model_helper.TensorProtosDBInput
Summary:
Caffe2 user was confused when model.TensorProtosDBINput([reader]) did not work. This is because of this outdated model helper function, that ignored the input blobs.
Added assertion to enforce correct usage. I did not want to make this work with reader input as well, since this probably should not be used anyway.

Reviewed By: amanrajdce

Differential Revision: D6380326

fbshipit-source-id: 6a50c2861f7f58c06cbfe3e86bde0f17a2b443cb