Make timing based test more likely to pass
This assumed that the expect statement would run within 1us, whereas
we only care it runs in less than the 100ms to check that it got reset.

Also upon construction, it may be so fast to be equal to 0
nanoseconds, so we add a sleep to force it to be non-zero.