Import the FindCUDA module from CMake
This addresses a problem where a generator expression for include
directories was passed to NVCC, where the generator expression itself
was prefixed by -I. As the NNPACK include directory generator
expression expands to multiple directories, the second and later ones
were not prefixed by -I, causing nvcc to return an error.

This was first fixed in CMake 3.7 (see Kitware/CMake@7ded655f).

To fix this, we can import and track FindCUDA from CMake upstream.

The most recent change to these files is Kitware/CMake@21806e6f.

This fixes #1397.