Make docker image built on Jenkins usable out of the box
The install prefix we use in our builds is /usr/local/caffe2. This is
not standard, so in order to load caffe2 from Python, the Python
interpreter must know where to find it. In a post-build section in the
Jenkins build script we know add a symlink to Python's dist-packages
directory and instruct the loader to look in /usr/local/caffe2/lib.
Together, these tricks make it usable out of the box.