Checking performance flags during init.
Adds 2 features:
(1) In cmake, allow the use of -march=native
(2) During initialization, check if Caffe2 is built with matching cpu
features of the current machine.

This helps us guarding performance claims in case the Caffe2 baseline is
built with limited computation capability.

Currently only added avx, avx2 and fma which are common.