[Dper2] Add gradient clipping
This diff is to support gradient clipping.
It is added as a net modifier after the gradients are computed.
See D6935400 for the abstraction.

Besides using a hard-threshold to clip the gradients,
we also support using the parameter norm to determine the
scaling ratio. This is similar to the behavior of LARS.